Lua: Add package.core, remove import_core()
