+++ data/libs/Player.lua
@@ -1,7 +1,7 @@
 -- Copyright © 2008-2020 Pioneer Developers. See AUTHORS.txt for details
 -- Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
-local Player = import_core("Player")
+local Player = package.core["Player"]
 local Serializer = import("Serializer")
 local Event = import("Event")
 local Game = import("Game")
