+++ src/lua/core/CoreFwdDecl.h
@@ -8,6 +8,9 @@
 // Forward-declared from ../LuaUtils.cpp
 int luaopen_utils(lua_State *L);
 
+// Set up the import() functionality
+int luaopen_import(lua_State *L);
+
 // Internal functions corresponding to pi_lua_import, but which leave the error message on the stack for Lua
 bool lua_import_core(lua_State *L, const std::string &importName);
 bool lua_import(lua_State *L, const std::string &importName, bool isFullName = false);
